+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Primary data class parameter cleanup [#1459](https://github.com/ie3-institute/simona/issues/1459)
 - Set Institute as first author within `CITATION.cff` [#1427](https://github.com/ie3-institute/simona/issues/1427)
 - Update Developers Guide - Release Process [#1466](https://github.com/ie3-institute/simona/issues/1466)
+- Made powerflow parameters in config options [#1443](https://github.com/ie3-institute/simona/issues/1443)
 
 ### Fixed
 - Fixes in Documentation, ScalaDocs, Code Style and more [#1397](https://github.com/ie3-institute/simona/issues/1397)